chore: Use Unref instead of Reset to avoid SEGFAULT
Since Node.js 20, the test `should invoke *all* callbacks from
different weak references` fails because of a SEGFAULT. After
debugging, it seems that the issue is that, after the first
Immediate Callback and the corresponding Reset call of ObjectInfo,
the destructor to this is called, invalidating the reference, and
SetImmediate accesses to invalid memory.

By using Unref, we ensure that `this` is only destructed after all
callbacks are actually called. This matches the expectations of
the constructor, where we trigger a new Ref().